+++ src/services/config/config.defaults.js
@@ -178,6 +178,8 @@ export class ConfigSetup {
           camera.motionTimeout =
             camera.motionTimeout === undefined || !(camera.motionTimeout >= 0) ? 15 : camera.motionTimeout;
 
+          camera.motionDelay = camera.motionDelay && camera.motionDelay <= 10 ? camera.motionDelay : undefined;
+
           // validate prebufferLength
           camera.prebufferLength =
             camera.prebufferLength >= 4 && camera.prebufferLength <= 8 ? camera.prebufferLength : 4;