+++ ui/src/views/Settings/subpages/account.vue
@@ -1,50 +1,51 @@
 <template lang="pug">
-.tw-w-full.tw-mt-8
+.tw-w-full
   v-progress-linear.loader(:active="loading" :indeterminate="loading" fixed top color="var(--cui-primary)")
 
   v-btn.save-btn(:class="fabAbove ? 'save-btn-top' : ''" v-scroll="onScroll" v-show="fab" color="success" transition="fade-transition" width="40" height="40" fab dark fixed bottom right @click="save" :loading="loadingProgress")
     v-icon {{ icons['mdiCheckBold'] }}
 
-  .tw-flex.tw-justify-between
-    .tw-block
-      .page-subtitle {{ $t('profile') }}
-      .page-subtitle-info {{ $t('general_information') }}
-    .image-upload.tw-ml-auto
-      label.profile-avatar-bg(for='file-input')
-        v-img.profile-avatar(v-on:error="handleErrorImg" :src="avatarSrc" alt="Avatar" width="4rem" height="4rem" style="border: 1px solid #1a1a1a")
-          template(v-slot:placeholder)
-            .tw-flex.tw-justify-center.tw-items-center.tw-h-full
-              v-progress-circular(indeterminate color="var(--cui-primary)" size="22")
-      input#file-input(type="file", name="photo", placeholder="Photo", required="", accept="image/png,image/jpeg", @change="changeProfileImg")
+  .tw-mb-7.tw-mt-5(v-if="!loading")
+    .tw-flex.tw-justify-between
+      .tw-block
+        .page-subtitle {{ $t('profile') }}
+        .page-subtitle-info {{ $t('general_information') }}
+      .image-upload.tw-ml-auto
+        label.profile-avatar-bg(for='file-input')
+          v-img.profile-avatar(v-on:error="handleErrorImg" :src="avatarSrc" alt="Avatar" width="4rem" height="4rem" style="border: 1px solid #1a1a1a")
+            template(v-slot:placeholder)
+              .tw-flex.tw-justify-center.tw-items-center.tw-h-full
+                v-progress-circular(indeterminate color="var(--cui-primary)" size="22")
+        input#file-input(type="file", name="photo", placeholder="Photo", required="", accept="image/png,image/jpeg", @change="changeProfileImg")
 
-  v-form.tw-w-full.tw-mt-4.tw-mb-8(ref="form" v-model="valid" lazy-validation)
-    label.form-input-label {{ $t('username') }}
-    v-text-field(v-model="form.username" :label="$t('username')" prepend-inner-icon="mdi-account" background-color="var(--cui-bg-card)" color="var(--cui-text-default)" :rules="rules.username" required solo)
-      template(v-slot:prepend-inner)
-        v-icon.text-muted {{ icons['mdiAccount'] }}
+    v-form.tw-w-full.tw-mt-4.tw-mb-8(ref="form" v-model="valid" lazy-validation)
+      label.form-input-label {{ $t('username') }}
+      v-text-field(v-model="form.username" :label="$t('username')" prepend-inner-icon="mdi-account" background-color="var(--cui-bg-card)" color="var(--cui-text-default)" :rules="rules.username" required solo)
+        template(v-slot:prepend-inner)
+          v-icon.text-muted {{ icons['mdiAccount'] }}
 
-    label.form-input-label {{ $t('signout_after') }}
-    v-select(ref="sessionTimer" :suffix="$t('hours')" :value="form.sessionTimer/3600 > 25 ? $t('never') : form.sessionTimer/3600" :items="sessionTimerSelect" prepend-inner-icon="mdi-timelapse" background-color="var(--cui-bg-card)" required solo)
-      template(v-slot:prepend-inner)
-        v-icon.text-muted {{ icons['mdiTimelapse'] }}
+      label.form-input-label {{ $t('signout_after') }}
+      v-select(ref="sessionTimer" :suffix="$t('hours')" :value="form.sessionTimer/3600 > 25 ? $t('never') : form.sessionTimer/3600" :items="sessionTimerSelect" prepend-inner-icon="mdi-timelapse" background-color="var(--cui-bg-card)" required solo)
+        template(v-slot:prepend-inner)
+          v-icon.text-muted {{ icons['mdiTimelapse'] }}
 
-    label.form-input-label {{ $t('permissions') }}: 
-    v-chip.tw-text-white.tw-ml-1(small v-for="perm in currentUser.permissionLevel" :key="perm" color="var(--cui-primary)") {{ perm }}
-    
-    v-divider.tw-my-8
-    
-    .page-subtitle.tw-mt-8 {{ $t('password') }}
-    .page-subtitle-info.tw-mb-8 {{ $t('change_your_password') }}
-    
-    label.form-input-label {{ $t('new_password') }}
-    v-text-field(v-model="form.password" label="******" autocomplete="new-password" :type="showNewPassword ? 'text' : 'password'" :append-icon="showNewPassword ? icons['mdiEye'] : icons['mdiEyeOff']" @click:append="showNewPassword = !showNewPassword" prepend-inner-icon="mdi-key-variant" background-color="var(--cui-bg-card)" color="var(--cui-text-default)" required solo)
-      template(v-slot:prepend-inner)
-        v-icon.text-muted {{ icons['mdiKeyVariant'] }}
+      label.form-input-label {{ $t('permissions') }}: 
+      v-chip.tw-text-white.tw-ml-1(small v-for="perm in currentUser.permissionLevel" :key="perm" color="var(--cui-primary)") {{ perm }}
+      
+      v-divider.tw-my-8
+      
+      .page-subtitle.tw-mt-8 {{ $t('password') }}
+      .page-subtitle-info.tw-mb-8 {{ $t('change_your_password') }}
+      
+      label.form-input-label {{ $t('new_password') }}
+      v-text-field(v-model="form.password" label="******" autocomplete="new-password" :type="showNewPassword ? 'text' : 'password'" :append-icon="showNewPassword ? icons['mdiEye'] : icons['mdiEyeOff']" @click:append="showNewPassword = !showNewPassword" prepend-inner-icon="mdi-key-variant" background-color="var(--cui-bg-card)" color="var(--cui-text-default)" required solo)
+        template(v-slot:prepend-inner)
+          v-icon.text-muted {{ icons['mdiKeyVariant'] }}
 
-    label.form-input-label {{ $t('new_password_verify') }}
-    v-text-field(v-model="form.password2" label="******" autocomplete="new-password-confirm" :type="showNewPasswordConfirm ? 'text' : 'password'" :append-icon="showNewPasswordConfirm ? icons['mdiEye'] : icons['mdiEyeOff']" @click:append="showNewPasswordConfirm = !showNewPasswordConfirm" prepend-inner-icon="mdi-key-variant" background-color="var(--cui-bg-card)" color="var(--cui-text-default)" :rules="rules.newpassword2" required solo)
-      template(v-slot:prepend-inner)
-        v-icon.text-muted {{ icons['mdiKeyVariant'] }}
+      label.form-input-label {{ $t('new_password_verify') }}
+      v-text-field(v-model="form.password2" label="******" autocomplete="new-password-confirm" :type="showNewPasswordConfirm ? 'text' : 'password'" :append-icon="showNewPasswordConfirm ? icons['mdiEye'] : icons['mdiEyeOff']" @click:append="showNewPasswordConfirm = !showNewPasswordConfirm" prepend-inner-icon="mdi-key-variant" background-color="var(--cui-bg-card)" color="var(--cui-text-default)" :rules="rules.newpassword2" required solo)
+        template(v-slot:prepend-inner)
+          v-icon.text-muted {{ icons['mdiKeyVariant'] }}
 </template>
 
 <script>