+++ ui/src/views/Settings/subpages/rekognition.vue
@@ -1,8 +1,8 @@
 <template lang="pug">
-.tw-w-full.tw-mt-8
+.tw-w-full
   v-progress-linear.loader(:active="loadingProgress" :indeterminate="loadingProgress" fixed top color="var(--cui-primary)")
 
-  .tw-mb-7(v-if="!loading")
+  .tw-mb-7.tw-mt-5(v-if="!loading")
     .page-subtitle {{ $t('aws') }}
     .page-subtitle-info {{ $t('amazon_rekognition') }}